+++ test_settings.py
@@ -68,6 +68,7 @@ INSTALLED_APPS = (
     'allauth.socialaccount.providers.daum',
     'allauth.socialaccount.providers.digitalocean',
     'allauth.socialaccount.providers.discord',
+    'allauth.socialaccount.providers.disqus',
     'allauth.socialaccount.providers.douban',
     'allauth.socialaccount.providers.doximity',
     'allauth.socialaccount.providers.draugiem',