+++ allauth/account/views.py
@@ -521,6 +521,7 @@ class EmailView(AjaxCapableProcessFormViewMixin, FormView):
         data = []
         for emailaddress in self.request.user.emailaddress_set.all():
             data.append({
+                'id': emailaddress.pk,
                 'email': emailaddress.email,
                 'verified': emailaddress.verified,
                 'primary': emailaddress.primary,