+++ ChangeLog.rst
@@ -1,7 +1,13 @@
 0.37.0 (unreleased)
 *******************
 
+Note worthy changes
+-------------------
+
 - The Battle.net login backend now recognizes ``apac`` as a valid region.
+
+- User model using a ``UUIDField`` as it's primary key can now be logged
+  in upon email confirmation (if using ``ACCOUNT_LOGIN_ON_EMAIL_CONFIRMATION``)
 
 
 0.36.0 (2018-05-08)